backendProject/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.4.0
django-rest-framework-nested==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.3.2
pytest-django==4.8.0
sqlparse==0.5.0
tzdata==2024.1
gunicorn
psycopg2-binary
python-dotenv
django-redis
django-filter==21.1
google-auth
requests

